Report the elapsed time more logically
Rather than look at the previous match, look at the next match.
This means that the labels can be set for when a "stage" of the build begins and will be correct as long as there is a label for when that "stage" ends.
The last timestamp for the build is used on the last match.